refactor server action as a functon to act as a wrapper calling a method from a class
