+++ src/modules/ideation/application/usecases/addIdeationUseCase.ts
@@ -3,10 +3,9 @@ import { TYPES } from "@/di/types";
 import { type IdeationApiPort } from "@/modules/ideation/ports/secondary/ideationApiPort";
 import { type AddIdeationRequestDto } from "@/modules/ideation/application/dtos/request.dto";
 import { type AddIdeationResponseDto } from "@/modules/ideation/application/dtos/response.dto";
-import { type AddIdeationPort } from "@/modules/ideation/ports/primary/addIdeationPort";
 
 @injectable()
-export class AddIdeationUseCase implements AddIdeationPort {
+export class AddIdeationUseCase {
   constructor(
     @inject(TYPES.IdeationApiPort)
     private readonly ideationApi: IdeationApiPort,
