+++ src/modules/ideation/application/usecases/addIdeationVoteUseCase.ts
@@ -3,9 +3,8 @@ import { TYPES } from "@/di/types";
 import { type IdeationVoteRequestDto } from "@/modules/ideation/application/dtos/request.dto";
 import { type IdeationVoteResponseDto } from "@/modules/ideation/application/dtos/response.dto";
 import { type IdeationApiPort } from "@/modules/ideation/ports/secondary/ideationApiPort";
-import { type AddIdeationVotePort } from "@/modules/ideation/ports/primary/addIdeationVotePort";
 
-export class AddIdeationVoteUseCase implements AddIdeationVotePort {
+export class AddIdeationVoteUseCase {
   constructor(
     @inject(TYPES.IdeationApiPort)
     private readonly ideationApi: IdeationApiPort,
