+++ src/modules/ideation/application/usecases/deleteIdeationUseCase.ts
@@ -3,9 +3,8 @@ import { TYPES } from "@/di/types";
 import { type DeleteIdeationRequestDto } from "@/modules/ideation/application/dtos/request.dto";
 import { type DeleteIdeationResponseDto } from "@/modules/ideation/application/dtos/response.dto";
 import { type IdeationApiPort } from "@/modules/ideation/ports/secondary/ideationApiPort";
-import { type DeleteIdeationPort } from "@/modules/ideation/ports/primary/deleteIdeationPort";
 
-export class DeleteIdeationUseCase implements DeleteIdeationPort {
+export class DeleteIdeationUseCase {
   constructor(
     @inject(TYPES.IdeationApiPort)
     private readonly ideationApi: IdeationApiPort,
