+++ src/modules/ideation/application/usecases/finalizeIdeationUseCase.ts
@@ -3,9 +3,8 @@ import { TYPES } from "@/di/types";
 import { type IdeationApiPort } from "@/modules/ideation/ports/secondary/ideationApiPort";
 import { type FinalizeIdeationRequestDto } from "@/modules/ideation/application/dtos/request.dto";
 import { type FinalizeIdeationResponseDto } from "@/modules/ideation/application/dtos/response.dto";
-import { type FinalizeIdeationPort } from "@/modules/ideation/ports/primary/finalizeIdeationPort";
 
-export class FinalizeIdeationUseCase implements FinalizeIdeationPort {
+export class FinalizeIdeationUseCase {
   constructor(
     @inject(TYPES.IdeationApiPort)
     private readonly ideationApi: IdeationApiPort,
