+++ src/modules/ideation/application/usecases/removeIdeationVoteUseCase.ts
@@ -3,9 +3,8 @@ import { TYPES } from "@/di/types";
 import { type IdeationApiPort } from "@/modules/ideation/ports/secondary/ideationApiPort";
 import { type IdeationVoteRequestDto } from "@/modules/ideation/application/dtos/request.dto";
 import { type IdeationVoteResponseDto } from "@/modules/ideation/application/dtos/response.dto";
-import { type RemoveIdeationVotePort } from "@/modules/ideation/ports/primary/removeIdeationVotePort";
 
-export class RemoveIdeationVoteUseCase implements RemoveIdeationVotePort {
+export class RemoveIdeationVoteUseCase {
   constructor(
     @inject(TYPES.IdeationApiPort)
     private readonly ideationApi: IdeationApiPort,
